+++ bioblend/_tests/TestGalaxyTools.py
@@ -130,8 +130,6 @@ class TestGalaxyTools(GalaxyTestBase.GalaxyTestBase):
         # TODO: Wait for results and verify it has 3 lines - 1 2 3, 4 5 6,
         # and 7 8 9.
 
-    # This test doesn't work any more on Galaxy 16.10 because that release uses an old Conda 3.19.3
-    @test_util.skip_unless_galaxy('release_17.01')
     def test_tool_dependency_install(self):
         installed_dependencies = self.gi.tools.install_dependencies('CONVERTER_fasta_to_bowtie_color_index')
         self.assertTrue(any(True for d in installed_dependencies if d.get('name') == 'bowtie' and d.get('dependency_type') == 'conda'), "installed_dependencies is %s" % installed_dependencies)