+++ bioblend/galaxy/histories/__init__.py
@@ -156,12 +156,6 @@ class HistoryClient(Client):
             For the purge option to work, the Galaxy instance must have the
             ``allow_user_dataset_purge`` option set to ``true`` in the
             ``config/galaxy.yml`` configuration file.
-
-        .. warning::
-            If you purge a dataset which has not been previously deleted,
-            Galaxy from release_15.03 to release_17.01 does not set the
-            ``deleted`` attribute of the dataset to ``True``, see
-            https://github.com/galaxyproject/galaxy/issues/3548
         """
         url = '/'.join((self._make_url(history_id, contents=True), dataset_id))
         payload = {}