+++ run_galaxy.sh
@@ -14,63 +14,7 @@ then
     . $GALAXY_LOCAL_ENV_FILE
 fi
 
-if [ -f scripts/common_startup.sh ]; then
-    ./scripts/common_startup.sh || exit 1
-else
-    if [ -f scripts/copy_sample_files.sh ]; then
-        ./scripts/copy_sample_files.sh
-    else
-        SAMPLES="
-            community_wsgi.ini.sample
-            datatypes_conf.xml.sample
-            external_service_types_conf.xml.sample
-            migrated_tools_conf.xml.sample
-            reports_wsgi.ini.sample
-            shed_tool_conf.xml.sample
-            tool_conf.xml.sample
-            shed_tool_data_table_conf.xml.sample
-            tool_data_table_conf.xml.sample
-            tool_sheds_conf.xml.sample
-            data_manager_conf.xml.sample
-            shed_data_manager_conf.xml.sample
-            openid_conf.xml.sample
-            universe_wsgi.ini.sample
-            tool-data/shared/ncbi/builds.txt.sample
-            tool-data/shared/ensembl/builds.txt.sample
-            tool-data/shared/ucsc/builds.txt.sample
-            tool-data/shared/ucsc/publicbuilds.txt.sample
-            tool-data/shared/ucsc/ucsc_build_sites.txt.sample
-            tool-data/shared/igv/igv_build_sites.txt.sample
-            tool-data/shared/rviewer/rviewer_build_sites.txt.sample
-            tool-data/*.sample
-            static/welcome.html.sample
-        "
-        # Create any missing config/location files
-        for sample in $SAMPLES; do
-            file=${sample%.sample}
-            if [ ! -f "$file" -a -f "$sample" ]; then
-                echo "Initializing $file from $(basename "$sample")"
-                cp "$sample" "$file"
-            fi
-        done
-    fi
-    # explicitly attempt to fetch eggs before running
-    FETCH_EGGS=1
-    for arg in "$@"; do
-        [ "$arg" = "--stop-daemon" ] && FETCH_EGGS=0; break
-    done
-    if [ $FETCH_EGGS -eq 1 ]; then
-        if ! python ./scripts/check_eggs.py -q; then
-            echo "Some eggs are out of date, attempting to fetch..."
-            if python ./scripts/fetch_eggs.py; then
-                echo "Fetch successful."
-            else
-                echo "Fetch failed."
-                exit 1
-            fi
-        fi
-    fi
-fi
+./scripts/common_startup.sh || exit 1
 
 # If there is a .venv/ directory, assume it contains a virtualenv that we
 # should run this instance in.
@@ -81,10 +25,6 @@ then
 fi
 
 python ./scripts/check_python.py || exit 1
-
-if [ -n "$GALAXY_UNIVERSE_CONFIG_DIR" ]; then
-    python ./scripts/build_universe_config.py "$GALAXY_UNIVERSE_CONFIG_DIR"
-fi
 
 if [ -z "$GALAXY_CONFIG_FILE" ]; then
     if [ -f universe_wsgi.ini ]; then